webapp/backend/requirements.txt
# CA Lobby API Dependencies - Phase 1.3a Backend Foundation
# Based on Phase 1.1 established packages and patterns

# Core Flask Framework
flask>=2.3.0
flask-cors>=4.0.0

# Database (BigQuery integration from Phase 1.1)
google-cloud-bigquery>=3.11.0
google-auth>=2.20.0

# Environment and Configuration
python-dotenv>=1.0.0

# Utilities (consistent with Phase 1.1)
pandas>=2.0.0
requests>=2.31.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Production WSGI Server
gunicorn>=21.2.0